changed name of `server` pkg to less generic `h5serv`
